reference note by uid
this simplifies the handling of notes massivily and eliminates race
conditions on creating notes.

the introduction of uuid makes the handling of new vs. update notes
obsolete. now we generate on the client a valid uuid which will be used
as the primary key. this way we don't need to update the client side
note with the actual database id after create.
furthermore, we now only have one 'update' action which handles
both creation and updates.