Date.as_json returns a different format in rails 6
The previous format was:
`Mon, 01 Aug 2016 15:33:00 UTC +00:00`

The new format is:
`2016-08-01T15:33:00.000Z`

The Note.as_json generated the date probably with `to_s`, which resulted
in the human readable format. With rails 6 it seems to use `as_json` on
the date property, which results in the ISO format.

I checked that in JavaScript `new Date(...)` returns the same date for
both formats:

> new Date("Mon, 01 Aug 2016 15:33:00 UTC +00:00")
  # Date Mon Aug 01 2016 17:33:00 GMT+0200 (Central European Summer Time)

> new Date("2016-08-01T15:33:00.000Z")
  # Date Mon Aug 01 2016 17:33:00 GMT+0200 (Central European Summer Time)

This is the logic that is used to create the dates in `Note.js`.

So, both formats seem to produce the same result.