operate NoteTaskPopulator on text_content
the note spec 'duplicates the images' started failing when setting
`#content`, because nokogiri tried to parse the base64 images.
while this may only be an issue with the test, it's a good pointer to
actually only operate on `#text_content` which is the database column.
`#content` is the attribute that CarrierWave::Base64ImageContent
operates on. if we modify that field we invoke the image parsing, which
may be a bit more unnecessary overhead.