load note on browser back
The browser back button, and especially the back button on mobile (in
the PWA) did change the url so far, but didn't show the note.
The note is now properly loaded on browser back.

For this to work the loaded notes are stored in a svelte store.
When navigating back, the note is tried in 3 different places:

1. From the sync storage, in case there are locale changes
2. From the $notes svelte store
3. From the server