prevent cursor jump in content on title change
When the title is changed and the user edits the content before the save
cycle, the cursor jumps to the beginning of the content after saving.

`shouldRerender` was previously explicitly set to `true` on title
change. I'm can't recall / reproduce why it was that way. Let's try with
no rerender and hope for the best.