fix share by not creating multiple notes
The current handling of initializing the `NoteEdit` component by relying
on `#beforeUpdate` is flawed. That function gets triggered too many
times. Every change to a variable that affects the component or any
child component will fire that event. To initialize from `share-target`
this is especially problematic, since it will create multiple notes from
the provided data.

We now change the whole thing to rely on reactive statements instead.
This is much more robust since we can rely on those triggering only when
the specific values have changed.