disable multiple tabs warning for PWA
When killing an app on mobile (by swiping it away) the `beforeunload`
event is not called, and therefore the tab counter is not decreased.
This leads to the warning to show up every time after that when starting
the app.

Sidenote: The autosave warning (on unsaved changes) doesn't work either.

The detection code is taken from
https://stackoverflow.com/a/52695341/232838.

Fixes https://github.com/koffeinfrei/mykonote/issues/538