focus title on new note
When clicking "+" or using the keyboard shortcut to create a new note
the code that sets the focus wasn't triggered anymore in case the
current note was already a new note. The note didn't change, and so the
focus code assumed to not do anything. By instantiating a new note we
now force focusing on the title.