fix logging
This could be releated to the rails 7.1. upgrade
https://github.com/koffeinfrei/unnote/pull/587. Somehow we lost the file
logging on production. But it could very well be that it was broken
before that. The < 7.1 logger definition was still there. I now added
the default 7.1. commented out. The default logger is the file logger to
`log/<env>.log`, which is what we want.

This patch also amends some configuration updates from the update tool
that I messed up.