show default note images in list again
With the recent npm packages update something broke related to imported
images. In `NotePicture.svelte` the two images `noteDefault` and
`taskDefault` would not be set as background images. In dev all is fine.
The issue seems to be with the build process. After some debugging I
decided to just disable inlining of images altogether, which fixes the
current issue. I can't be bothered with this kind of stupid problems.