replace secrets.yml by dotenv
With 944c8dc dotenv was dropped as a dependency of panter-deploy. We
still want to use it though. Since secrets.yml has long been deprecated
and doesn't work anymore with rails 7.2 we now replace setting up
`SECRET_KEY_BASE` with dotenv instead.
This issue was detected when trying to `cap deploy` and blowing up due
to the missing `SECRET_KEY_BASE`.